requirements.txt
alabaster==0.7.12
Babel==2.3.4
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
decorator==4.0.10
docutils==0.12
ebaysdk==2.1.5
feedparser==5.2.1
gevent==1.3.4
greenlet==0.4.13
html2text==2016.9.19
idna==2.7
imagesize==1.2.0
Jinja2==2.10.1
libsass==0.12.3
lxml==4.2.3
Mako==1.0.4
MarkupSafe==0.23
mock==2.0.0
num2words==0.5.6
ofxparse==0.16
packaging==20.1
passlib==1.6.5
pbr==5.4.4
Pillow==6.1.0
psutil==4.3.1
psycopg2==2.7.3.1
pydot==1.2.3
Pygments==2.5.2
PyJWT==1.7.1
pyldap==2.4.28
pyparsing==2.1.10
PyPDF2==1.26.0
pyserial==3.1.1
python-dateutil==2.5.3
python-stdnum==1.13
pytz==2016.7
pyusb==1.0.0
qrcode==5.3
reportlab==3.3.0
requests==2.20.0
simplejson==3.17.0
six==1.14.0
snowballstemmer==2.0.0
soupsieve==1.9.5
Sphinx==2.4.2
sphinx-patchqueue==1.0.4
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
suds-jurko==0.6
unidiff==0.5.5
urllib3==1.24.3
vatnumber==1.2
vobject==0.9.3
Werkzeug==0.11.15
xlrd==1.0.0
XlsxWriter==0.9.3
xlwt==1.3.0
genshi==0.7.3
py3o.template==0.10.0 
pyjon.utils==0.7
py3o.formats==0.3
